+++ pkg/manager/manager_test.go
@@ -1,7 +1,1 @@
 package manager
-
-import "testing"
-
-func TestSymlinking(t *testing.T) {
-
-}
